[refactor]: 전역 파라미터 검증 클래스로 변경
